refactor: 💡 create new directories
